+++ ta4j-core/src/main/java/org/ta4j/core/indicators/bollinger/BollingerBandFacade.java
@@ -53,12 +53,12 @@ public class BollingerBandFacade {
      * @param barCount  the number of periods used for the indicators
      * @param k         the multiplier used to calculate the upper and lower bands
      */
-    public BollingerBandFacade(BarSeries barSeries, int barCount, Number k) {
+    public BollingerBandFacade(final BarSeries barSeries, final int barCount, final Number k) {
         this.price = NumericIndicator.of(new ClosePriceIndicator(barSeries));
-        this.middle = NumericIndicator.of(price.sma(barCount));
-        final NumericIndicator stdev = price.stddev(barCount);
-        this.upper = middle.plus(stdev.multipliedBy(k));
-        this.lower = middle.minus(stdev.multipliedBy(k));
+        this.middle = NumericIndicator.of(this.price.sma(barCount));
+        final NumericIndicator stdev = this.price.stddev(barCount);
+        this.upper = this.middle.plus(stdev.multipliedBy(k));
+        this.lower = this.middle.minus(stdev.multipliedBy(k));
     }
 
     /**
@@ -68,12 +68,12 @@ public class BollingerBandFacade {
      * @param barCount  the number of periods used for the indicators
      * @param k         the multiplier used to calculate the upper and lower bands
      */
-    public BollingerBandFacade(Indicator<Num> indicator, int barCount, Number k) {
+    public BollingerBandFacade(final Indicator<Num> indicator, final int barCount, final Number k) {
         this.price = NumericIndicator.of(indicator);
-        this.middle = NumericIndicator.of(price.sma(barCount));
-        final NumericIndicator stdev = price.stddev(barCount);
-        this.upper = middle.plus(stdev.multipliedBy(k));
-        this.lower = middle.minus(stdev.multipliedBy(k));
+        this.middle = NumericIndicator.of(this.price.sma(barCount));
+        final NumericIndicator stdev = this.price.stddev(barCount);
+        this.upper = this.middle.plus(stdev.multipliedBy(k));
+        this.lower = this.middle.minus(stdev.multipliedBy(k));
     }
 
     /**
@@ -83,7 +83,7 @@ public class BollingerBandFacade {
      *         price.
      */
     public NumericIndicator middle() {
-        return middle;
+        return this.middle;
     }
 
     /**
@@ -93,7 +93,7 @@ public class BollingerBandFacade {
      *         standard deviation.
      */
     public NumericIndicator upper() {
-        return upper;
+        return this.upper;
     }
 
     /**
@@ -103,7 +103,7 @@ public class BollingerBandFacade {
      *         multiple of standard deviation.
      */
     public NumericIndicator lower() {
-        return lower;
+        return this.lower;
     }
 
     /**
@@ -113,7 +113,7 @@ public class BollingerBandFacade {
      *         middle
      */
     public NumericIndicator bandwidth() {
-        return upper.minus(lower).dividedBy(middle).multipliedBy(100);
+        return this.upper.minus(this.lower).dividedBy(this.middle).multipliedBy(100);
     }
 
     /**
@@ -122,7 +122,7 @@ public class BollingerBandFacade {
      * @return an object that calculates %B from close price, BB upper and lower
      */
     public NumericIndicator percentB() {
-        return price.minus(lower).dividedBy(upper.minus(lower));
+        return this.price.minus(this.lower).dividedBy(this.upper.minus(this.lower));
     }
 
 }