+++ ta4j-core/src/test/java/org/ta4j/core/indicators/bollinger/PercentBIndicatorTest.java
@@ -39,42 +39,41 @@ public class PercentBIndicatorTest extends AbstractIndicatorTest<Indicator<Num>,
 
     private ClosePriceIndicator closePrice;
 
-    public PercentBIndicatorTest(NumFactory numFactory) {
+    public PercentBIndicatorTest(final NumFactory numFactory) {
         super(numFactory);
     }
 
     @Before
     public void setUp() {
-        var data = new MockBarSeriesBuilder().withNumFactory(numFactory)
+        final var data = new MockBarSeriesBuilder().withNumFactory(this.numFactory)
                 .withData(10, 12, 15, 14, 17, 20, 21, 20, 20, 19, 20, 17, 12, 12, 9, 8, 9, 10, 9, 10)
                 .build();
-        closePrice = new ClosePriceIndicator(data);
+        this.closePrice = new ClosePriceIndicator(data);
     }
 
     @Test
     public void percentBUsingSMAAndStandardDeviation() {
-
-        var pcb = new PercentBIndicator(closePrice, 5, 2);
+        final var pcb = new PercentBIndicator(this.closePrice, 5, 2);
 
         assertTrue(pcb.getValue(0).isNaN());
-        assertNumEquals(0.75, pcb.getValue(1));
-        assertNumEquals(0.8244, pcb.getValue(2));
-        assertNumEquals(0.6627, pcb.getValue(3));
-        assertNumEquals(0.8517, pcb.getValue(4));
-        assertNumEquals(0.90328, pcb.getValue(5));
-        assertNumEquals(0.83, pcb.getValue(6));
-        assertNumEquals(0.6552, pcb.getValue(7));
-        assertNumEquals(0.5737, pcb.getValue(8));
-        assertNumEquals(0.1047, pcb.getValue(9));
-        assertNumEquals(0.5, pcb.getValue(10));
-        assertNumEquals(0.0284, pcb.getValue(11));
-        assertNumEquals(0.0344, pcb.getValue(12));
-        assertNumEquals(0.2064, pcb.getValue(13));
-        assertNumEquals(0.1835, pcb.getValue(14));
-        assertNumEquals(0.2131, pcb.getValue(15));
-        assertNumEquals(0.3506, pcb.getValue(16));
-        assertNumEquals(0.5737, pcb.getValue(17));
-        assertNumEquals(0.5, pcb.getValue(18));
-        assertNumEquals(0.7673, pcb.getValue(19));
+        assertNumEquals(0.6768, pcb.getValue(1)); // [10,12]
+        assertNumEquals(0.7649, pcb.getValue(2)); // [10,12,15]
+        assertNumEquals(0.6409, pcb.getValue(3)); // [10,12,15,14]
+        assertNumEquals(0.8146, pcb.getValue(4)); // [10,12,15,14,17]
+        assertNumEquals(0.8607, pcb.getValue(5)); // [12,15,14,17,20]
+        assertNumEquals(0.7951, pcb.getValue(6)); // [15,14,17,20,21]
+        assertNumEquals(0.6388, pcb.getValue(7)); // [14,17,20,21,20]
+        assertNumEquals(0.5659, pcb.getValue(8)); // [17,20,21,20,20]
+        assertNumEquals(0.1464, pcb.getValue(9)); // [20,21,20,20,19]
+        assertNumEquals(0.5000, pcb.getValue(10)); // [21,20,20,19,20]
+        assertNumEquals(0.0782, pcb.getValue(11)); // [20,20,19,20,17]
+        assertNumEquals(0.0835, pcb.getValue(12)); // [20,19,20,17,12]
+        assertNumEquals(0.2373, pcb.getValue(13)); // [19,20,17,12,12]
+        assertNumEquals(0.2169, pcb.getValue(14)); // [20,17,12,12,9]
+        assertNumEquals(0.2433, pcb.getValue(15)); // [17,12,12,9,8]
+        assertNumEquals(0.3663, pcb.getValue(16)); // [12,12,9,8,9]
+        assertNumEquals(0.5659, pcb.getValue(17)); // [12,9,8,9,10]
+        assertNumEquals(0.5000, pcb.getValue(18)); // [9,8,9,10,9]
+        assertNumEquals(0.7390, pcb.getValue(19)); // [8,9,10,9,10]
     }
 }